Start of torrent generator
This is the initial pass at generating metadata for files that get
dropped into the directory. This is implemented serially since the
intended use case is a long-running server which won't be adding lots of
files at any one time.

Parallel would certainly be faster in the start-up case, but there are
lots of variables re: consecutive reads vs. random reads that you'd get
by doing lots of parallel hashes. Another good option would be to cache
the metadata so we don't have to recalculate every launch.