Initial tracker implementation
This is very basic, but it works. This just gives you a list of peers
(first 50) and lets you download. This will obviously not work for any
real setup.

With this commit, Distributor had its first successful file
distribution. Seeding / Fetching are done by BT clients on the target
machines, with Distributor just handling the coordination.