Add event/num_want support
This allows clients to request however many peers they want, plus it
also supports the 'stopped' verb to remove the peer from the list of
valid peers.

This also introduces some IP collision logic which removes the older
peer when a newer peer with the same IP address starts up. This is not
strictly correct behavior, but it should suffice until I add expiration
of peers (when they get unhealthy).