Fix race condition around file metadata.
In another application where I'm using distributor as a library,
the tests failed with the library due to a data race condition
detected by the go test framework when run with:

    go test -p=1 -race -covermode atomic -coverprofile=/tmp/coverage.out MyPackage

Poking through the code, the race condition involved some threads reading
File's MetadataInfo field, while other fields might be writing it.

This change is a brute-forceish fix: it creates a new mutex, and wraps
every access to the MetadataInfo field in Lock/Unlock calls. Since
there's only one writer, and the reads never hold the lock for longer
than a single read of the field, this should be deadlock-free.

Using distributor with this check causes my tests on the other project
to run successfully - no more race detected!